Rename api/insert to certadd to be more consistent
