+++ api/certadd/insert_test.go
@@ -1,4 +1,4 @@
-package insert
+package certadd
 
 import (
 	"bytes"
